+++ src/main/java/io/github/seonrizee/scheduler/common/annotation/SessionUser.java
@@ -1,4 +1,4 @@
-package io.github.seonrizee.scheduler.auth;
+package io.github.seonrizee.scheduler.common.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,5 +7,5 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface SessionUserId {
+public @interface SessionUser {
 }
